% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{add_sessionid_to_db}
\alias{add_sessionid_to_db}
\title{add_sessionid_to_db}
\usage{
add_sessionid_to_db(user, sessionid, conn = db)
}
\value{
This function saves to your database.
}
\description{
This function must accept two parameters: user and sessionid. It will be called whenever the user
}
