% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{install_db_backend}
\alias{install_db_backend}
\title{Install Database Backend Package}
\usage{
install_db_backend(db_type)
}
\arguments{
\item{db_type}{A string specifying the database type.}
}
\value{
TRUE if installation was successful, FALSE otherwise.
}
\description{
Installs the appropriate DBI backend package based on the selected database type.
}
