% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{retrieve_credentials}
\alias{retrieve_credentials}
\title{Verify User Credentials from SQLite}
\usage{
retrieve_credentials(db_file)
}
\arguments{
\item{db_file}{Path to the SQLite database file.}

\item{user}{The username to verify.}

\item{password}{The password to verify.}
}
\value{
A list with user details if the password is correct. Throws an error if verification fails.
}
\description{
Retrieves and verifies the provided password against the stored hashed password in SQLite.
}
\examples{
\dontrun{
user_info <- verify_credentials("user", "password", "app_config.sqlite")
}
}
