% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R, R/utils_helpers.R
\name{style_my_workbook}
\alias{style_my_workbook}
\title{Style My Workbook}
\usage{
style_my_workbook(wb, dt, SheetName, alt_rows = TRUE)

style_my_workbook(wb, dt, SheetName, alt_rows = TRUE)
}
\arguments{
\item{wb}{workbook to apply the style}

\item{dt}{dataframe to put in the wb}

\item{SheetName}{name of the spreadsheet}

\item{alt_rows}{activate or not the color alternation}
}
\value{
a workbook

a workbook
}
\description{
Apply a style to a given workbook

Apply a style to a given workbook
}
\examples{
\dontrun{

options(openxlsx.datetimeFormat = "yyyy-mm-dd")
wb <- openxlsx::createWorkbook()
style_my_workbook(wb,dt, "TestStyle", TRUE)
openxlsx::saveWorkbook(wb, path, overwrite = TRUE)
}

\dontrun{

options(openxlsx.datetimeFormat = "yyyy-mm-dd")
wb <- openxlsx::createWorkbook()
style_my_workbook(wb,dt, "TestStyle", TRUE)
openxlsx::saveWorkbook(wb, path, overwrite = TRUE)
}

}
